create a separate ConfigFileParser for OSN configs so that an OSNStreamProcessor owns them and not a Consumer/ConsumerGroup
